+++ app/Modules/School/Models/School.php
@@ -2,11 +2,14 @@
 
 namespace App\Modules\School\Models;
 
+use App\Support\Concerns\HasTranslations;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class School extends Model
 {
+    use HasTranslations;
+
     protected $fillable = [
         'name',
         'institution_code',